+++ src/main/java/org/betonquest/betonquest/compatibility/citizens/CitizensParticle.java
@@ -5,7 +5,7 @@ import lombok.CustomLog;
 import net.citizensnpcs.api.CitizensAPI;
 import net.citizensnpcs.api.npc.NPC;
 import org.betonquest.betonquest.BetonQuest;
-import org.betonquest.betonquest.api.config.QuestPackage;
+import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.profiles.OnlineProfile;
 import org.betonquest.betonquest.compatibility.effectlib.EffectLibIntegrator;
 import org.betonquest.betonquest.compatibility.protocollib.hider.NPCHider;
@@ -63,7 +63,7 @@ public class CitizensParticle extends BukkitRunnable {
             // load the condition check interval
             interval = section.getInt("check_interval", 100);
             if (interval <= 0) {
-                LOG.warn(pack, "Could not load npc effects of package " + pack.getPackagePath() + ": " +
+                LOG.warn(pack, "Could not load npc effects of package " + pack.getQuestPath() + ": " +
                         "Check interval must be bigger than 0.");
                 enabled = false;
                 return;
@@ -89,7 +89,7 @@ public class CitizensParticle extends BukkitRunnable {
                 // load the interval between animations
                 effect.interval = settings.getInt("interval", 100);
                 if (effect.interval <= 0) {
-                    LOG.warn(pack, "Could not load npc effect " + key + " in package " + pack.getPackagePath() + ": " +
+                    LOG.warn(pack, "Could not load npc effect " + key + " in package " + pack.getQuestPath() + ": " +
                             "Effect interval must be bigger than 0.");
                     continue;
                 }