+++ src/main/java/org/betonquest/betonquest/compatibility/magic/WandCondition.java
@@ -54,7 +54,7 @@ public class WandCondition extends Condition {
                     final VariableNumber level;
                     final String[] spellParts = spell.split(":");
                     try {
-                        level = new VariableNumber(instruction.getPackage().getPackagePath(), spellParts[1]);
+                        level = new VariableNumber(instruction.getPackage().getQuestPath(), spellParts[1]);
                     } catch (final InstructionParseException e) {
                         throw new InstructionParseException("Could not parse spell level", e);
                     }