+++ src/main/java/org/betonquest/betonquest/compatibility/protocollib/conversation/MenuConvIO.java
@@ -15,7 +15,7 @@ import net.md_5.bungee.api.chat.BaseComponent;
 import net.md_5.bungee.api.chat.TextComponent;
 import org.apache.commons.lang3.StringUtils;
 import org.betonquest.betonquest.BetonQuest;
-import org.betonquest.betonquest.api.config.QuestPackage;
+import org.betonquest.betonquest.api.config.quest.QuestPackage;
 import org.betonquest.betonquest.api.profiles.OnlineProfile;
 import org.betonquest.betonquest.config.Config;
 import org.betonquest.betonquest.conversation.ChatConvIO;
@@ -141,7 +141,7 @@ public class MenuConvIO extends ChatConvIO {
                 }
             }
         } catch (final IllegalArgumentException e) {
-            LOG.warn(conv.getPackage(), conv.getPackage().getPackagePath() + ": Invalid data for 'control_cancel': " + configControlCancel, e);
+            LOG.warn(conv.getPackage(), conv.getPackage().getQuestPath() + ": Invalid data for 'control_cancel': " + configControlCancel, e);
         }
         try {
             for (final CONTROL control : Arrays.stream(configControlSelect.split(","))
@@ -153,7 +153,7 @@ public class MenuConvIO extends ChatConvIO {
                 }
             }
         } catch (final IllegalArgumentException e) {
-            LOG.warn(conv.getPackage(), conv.getPackage().getPackagePath() + ": Invalid data for 'control_select': " + configControlSelect, e);
+            LOG.warn(conv.getPackage(), conv.getPackage().getQuestPath() + ": Invalid data for 'control_select': " + configControlSelect, e);
         }
         try {
             for (final CONTROL control : Arrays.stream(configControlMove.split(","))
@@ -164,7 +164,7 @@ public class MenuConvIO extends ChatConvIO {
                 }
             }
         } catch (final IllegalArgumentException e) {
-            LOG.warn(conv.getPackage(), conv.getPackage().getPackagePath() + ": Invalid data for 'control_move': " + configControlMove, e);
+            LOG.warn(conv.getPackage(), conv.getPackage().getQuestPath() + ": Invalid data for 'control_move': " + configControlMove, e);
         }
     }