+++ src/main/java/org/betonquest/betonquest/events/CommandEvent.java
@@ -52,7 +52,7 @@ public class CommandEvent extends QuestEvent {
                         String com = command.command;
                         for (final String var : command.variables) {
                             com = com.replace(var, BetonQuest.getInstance().getVariableValue(
-                                    instruction.getPackage().getPackagePath(), var, onlineProfile));
+                                    instruction.getPackage().getQuestPath(), var, onlineProfile));
                         }
                         Bukkit.getServer().dispatchCommand(Bukkit.getConsoleSender(), com);
                     }
@@ -62,7 +62,7 @@ public class CommandEvent extends QuestEvent {
                         String com = command.command;
                         for (final String var : command.variables) {
                             com = com.replace(var, BetonQuest.getInstance().getVariableValue(
-                                    instruction.getPackage().getPackagePath(), var, profile));
+                                    instruction.getPackage().getQuestPath(), var, profile));
                         }
                         final String finalCom = com;
                         Bukkit.getScheduler().callSyncMethod(BetonQuest.getInstance(), () -> Bukkit.dispatchCommand(Bukkit.getConsoleSender(), finalCom));