+++ src/main/java/org/betonquest/betonquest/events/PartyEvent.java
@@ -34,7 +34,7 @@ public class PartyEvent extends QuestEvent {
     @Override
     protected Void execute(final Profile profile) throws QuestRuntimeException {
         final List<OnlineProfile> members = Utils.getParty(profile.getOnlineProfile().get(), range.getDouble(profile), instruction.getPackage()
-                .getPackagePath(), conditions);
+                .getQuestPath(), conditions);
         for (final OnlineProfile memberID : members) {
             for (final EventID event : events) {
                 BetonQuest.event(memberID, event);